+++ crashsimilarity/utils.py
@@ -8,6 +8,17 @@ from smart_open import smart_open
 from crashsimilarity.downloader import SocorroDownloader
 
 
+def utc_today():
+    return datetime.utcnow().date()
+
+
+def read_files(file_names, open_file_function=smart_open):
+    for name in file_names:
+        with open_file_function(name) as f:
+            for line in f:
+                yield line if isinstance(line, str) else line.decode('utf8')
+
+
 class StackTracesGetter(object):
     @staticmethod
     def get_stack_traces_for_signature(fnames, signature, traces_num=100):
@@ -57,37 +68,9 @@ class StackTraceProcessor(object):  # just a namespace, actually
                 yield (processed, data['signature'].lower())
 
 
-def utc_today():
-    return datetime.utcnow().date()
-
-
-def read_files(file_names, open_file_function=smart_open):
-    for name in file_names:
-        with open_file_function(name) as f:
-            for line in f:
-                yield line if isinstance(line, str) else line.decode('utf8')
-
-
-def delete_old_models(current_date, path, force_train):
-    """
-    Get list of trained models inside the models directory,
-    delete all models in case of user forces new training,
-    if not, delete all models except of today's model
-    """
-    if not os.path.isdir(path):
-        return
-
-    if force_train:
-        old_models = [model for model in os.listdir(path)]
-    else:
-        old_models = [model for model in os.listdir(path) if current_date not in model]
-    for model in old_models:
-        os.remove(path + model)
-
-
 def create_dir(path):
     try:
-        os.mkdir(path)
+        os.makedirs(path)
     except OSError as e:
         if e.errno != errno.EEXIST:
             raise e
